requirements.txt
asyncpg
aiogram
APScheduler
python-decouple
python-dotenv
psycopg